Added new functions to the Cell Complex class
