+++ pytspl/cell_complex/cell_complex.py
@@ -44,7 +44,7 @@ class CellComplex:
         Returns:
             SimplicialComplex: The resulting simplicial complex.
         """
-        from pytspl.simplicial_complex import SimplicialComplex  
+        from pytspl.simplicial_complex import SimplicialComplex
 
         # Filter only the polygons that are triangles
         triangles = [p for p in self.polygons if len(p) == 3]
@@ -93,5 +93,133 @@ class CellComplex:
 
         return B2
 
+    def generate_coordinates(self) -> dict:
+        """
+        Generate the coordinates of the nodes using spring layout
+        if the coordinates of the cc don't exist.
+
+        Returns:
+            dict: Coordinates of the nodes.
+        """
+        import networkx as nx
+
+        print("WARNING: No coordinates found.")
+        print("Generating coordinates using spring layout.")
+
+        G = nx.Graph()
+        G.add_nodes_from(self.nodes)
+        G.add_edges_from(self.edges)
+
+        coordinates = nx.spring_layout(G)
+        return coordinates
+
+    def tocsr(self, matrix: np.ndarray) -> csr_matrix:
+        """
+        Convert a numpy array to a csr_matrix.
+
+        Args:
+            matrix (np.ndarray): Numpy array to convert.
+
+        Returns:
+            csr_matrix: Compressed Sparse Row matrix.
+        """
+        return csr_matrix(matrix, dtype=float)
+    
+    def incidence_matrix(self, rank: int) -> csr_matrix:
+        """
+        Compute the incidence matrix of the simplicial complex.
+
+        Args:
+            rank (int): Rank of the incidence matrix.
+
+        Returns:
+            csr_matrix: Incidence matrix of the simplicial complex.
+        """
+        if rank == 0:
+            return np.ones(len(self.nodes), dtype=float)
+        elif rank == 1:
+            return self.tocsr(self.B1)
+        elif rank == 2:
+            return self.tocsr(self.B2)
+        else:
+            raise ValueError(
+                "Rank cannot be larger than the dimension of the complex."
+            )
+
+    def laplacian_matrix(self) -> csr_matrix:
+        """
+        Compute the Laplacian matrix of the cell complex.
+
+        Returns:
+            csr_matrix: Laplacian matrix of the cell complex.
+        """
+        B1 = self.tocsr(self.B1)
+        return B1 @ B1.T
+    
+    def lower_laplacian_matrix(self, rank: int = 1) -> csr_matrix:
+        """
+        Compute the lower Laplacian matrix of the cell complex.
+
+        Args:
+            rank (int): Rank of the lower Laplacian matrix.
+
+        ValueError:
+            If the rank is not 1 or 2.
+
+        Returns:
+            csr_matrix: Lower Laplacian matrix of the cell complex.
+        """
+        if rank == 1:
+            B1 = self.incidence_matrix(rank=1)
+            return B1.T @ B1
+        elif rank == 2:
+            B2 = self.incidence_matrix(rank=2)
+            return B2.T @ B2
+        else:
+            raise ValueError("Rank must be either 1 or 2.")
+        
+    def upper_laplacian_matrix(self, rank: int = 1) -> csr_matrix:
+        """
+        Compute the upper Laplacian matrix of the cell complex.
+
+        Args:
+            rank (int): Rank of the upper Laplacian matrix.
+
+        ValueError:
+            If the rank is not 0 or 1.
+
+        Returns:
+            csr_matrix: Upper Laplacian matrix of the cell complex.
+        """
+        if rank == 0:
+            return self.laplacian_matrix()
+        elif rank == 1:
+            B2 = self.incidence_matrix(rank=2)
+            return B2 @ B2.T
+        else:
+            raise ValueError("Rank must be either 0 or 1.")
+        
+    def hodge_laplacian_matrix(self, rank: int = 1) -> csr_matrix:
+        """
+        Compute the Hodge Laplacian matrix of the cell complex.
+
+        Args:
+            rank (int): Rank of the Hodge Laplacian matrix.
+
+        ValueError:
+            If the rank is not 0, 1, or 2.
+
+        Returns:
+            csr_matrix: Hodge Laplacian matrix of the cell complex.
+        """
+        if rank == 0:
+            return self.laplacian_matrix()
+        elif rank == 1:
+            return self.lower_laplacian_matrix(
+                rank=rank
+            ) + self.upper_laplacian_matrix(rank=rank)
+        else:
+            raise ValueError("Rank must be between 0 and 2.")
+
